코드 수정

+++ getframe_youtube.py
@@ -8,7 +8,7 @@ import cv2
 import os
 import re
 
-# 유튜브 영상 하나 당 생기는 클래스
+# 유튜브 url(영상) 하나에 대해 생기는 클래스
 class Youtube_to_image:
     def __init__(self, ss_array, output_path, url):
         self.ss_array = ss_array
@@ -17,41 +17,35 @@ class Youtube_to_image:
 
     def youtube_download(self):
         download_folder = os.path.join(self.output_path, 'video')
+
+        new_file = os.path.join(download_folder, self.url) + '.mp4'
+        # 이미 영상이 존재하는 경우
+        if os.path.exists(new_file):
+            print(self.url + ' video file already exists')
+            return new_file
+
         yt = YouTube('https://www.youtube.com/watch?v=' + self.url, use_oauth=True, allow_oauth_cache=True)
+
         try:
             video_name = yt.title
-        except Exception as e:
-            print(e)
-            return 0
-        
-        video_name = re.sub(r'[#:;.,\<"|/>*?]', '', video_name)
-        video_name = re.sub(r"[']", '', video_name)
 
-        try:
+            # 영상 title 로컬에 저장되는 형식으로 변경
+            video_name = re.sub(r'[#:;.,\<"|/>*?]', '', video_name)
+            video_name = re.sub(r"[']", '', video_name)
             stream = yt.streams.filter(file_extension='mp4').get_highest_resolution()
-        except Exception as e:
-            print(e)
-            return 0
-        
-        try:
             stream.download(download_folder)
-        except Exception as e:
-            print(e)
-            return 0
-        
-        dir = download_folder
-        old_file = os.path.join(dir, video_name) + '.mp4'
-        print(video_name)
-        new_file = os.path.join(dir, self.url) + '.mp4'
-        
-        try:
+
+            # 영상 title url로 형식으로 변경
+            old_file = os.path.join(download_folder, video_name) + '.mp4'
             os.rename(old_file, new_file)
+
         except Exception as e:
             print(e)
             return 0
+        
 
         video_path = new_file
-        print(self.url + ' image download : success')
+        print(self.url + ' video download : success')
         return video_path
 
 
@@ -68,12 +62,13 @@ class Youtube_to_image:
             process.communicate()
 
             if process.returncode == 0:
-                print(self.url + ' : success')
+                print(self.url + ' image download : success')
             else:
-                print(self.url + ' : failed')
+                print(self.url + ' image download : failed')
         
 
 def main():
+    # 추출된 이미지를 담을 output path 설정
     output_path = "C:/Users/Snucem_W1/Desktop/YoutubeToImage"
     
     # 구글 스프레드시트 연동
@@ -89,13 +84,15 @@ def main():
     doc = gc.open_by_url(spreadsheet_url)
 
     worksheet = doc.worksheet('Youtube')
+    # B, C, D열에 해당하는 정보만 가져오기
     array = np.array(worksheet.get_all_values()[2:])[:, 1:4]
 
     url_list = list(set(array[:, 0]))
-    print(len(url_list))
     url_list.remove('')
+    print('number of videos in table :', len(url_list))
 
     for url in url_list:
+        # 해당 url에 대응하는 start, stop 표를 가져오기
         ss_array = array[array[:, 0] == url][:, 1:3]
         yt = Youtube_to_image(ss_array, output_path, url)
         video_path = yt.youtube_download()
